Merging dev in 22-add-hamiltonian-calcs
I originally branched off of a previous branch for updating the equations of motion to polar form. The rebase of this branch got messy, and I decided to start of with the original source of this branch. I am merging the latest version of dev into this branch
to get the latest set of approved changes.

# Conflicts:
#	src/python/Propagation.py
#	src/python/Spacecraft.py
#	src/python/TwoBody_Orb2Orb_Transfer_Env.py
#	src/scripts/TBO_Transfer_Env_Test.py