Supplying heuristic initial guess for co-states
For the direct shooting method that will be used for solving the system of equations, we need to supply an initial guess for the costates. Based on my research, these should basically just be on the order of magnitudes for the state vectors themselves. Also setting lambda_r_0 and lambda_m_0 to negative numbers typically helps with convergence.