Adding solution finder method
This method wraps the functionality needed to find a solution to the TPBVP needed to solve for Hamiltonian control. The root finder method uses the initial conditions and attemps to solve for the lam_0 values that solve the TPBVP. The resultant solution is then propagated, and the solution object is returned for examination and plotting.