Reducing scope of test script for HC
The Hamiltonian control test script was reduced in scope down to basically initializing the Hamiltonian controller object and calling the "solution finder" method to run the logic implemented in the Hamiltonial control class. Until the issues with the solution finder are resolved the script will remain focused on just creating this controller object and calling the solution finder method.