Updating root finder method to handle smoothing
The root finder method is now wrapped in some logic that uses the smoothing parameter epsilon and also can do some convergence error handling with the root finder function. In cases where epsilon is approaching zero, the root finder method can struggle to find roots with very high accuracy when u is a bang-bang. The root finder tolerance is relaxed up to a point to allow some flexibility with the targeting process. If a max number of attempts is reached or the tolerance is increased past a pre-defined upper limit, the targeting procedure fails.