Removing heuristic success constraint on co-state solution.
I decided this was not a good idea to enforce this arbitrary constraint that the largest co-state be less than one. Previous thinking was that there is a chance for convergence on local optimum fuel mass trajectories that globally weren't optimal and used more mass than necessary. These suboptimal trajectories usually coincided with large co-state values. With other robust improvements, this has largely gone away, and I don't think it's a good idea to arbitrarily constrain the solution space since this large co-state suboptimality may not always be true.