Adding a training data generation file.
This includes two functions (and probably more later). generate_nn_training_data is meant to be used in multithreading, the function basically does all of the trajectory solution work and returns a flag if the solver converged, an ephemeris filename and the thread id. The ephem is written only if the solver converged.
The second method is basically a thread manager, it establishes the workers and the job queue and then prints the results as each thread completes its job.